requirements.txt
ddddocr==1.4.3
Pillow==9.1.0
pycryptodome
pycryptodomex==3.14.1
PyNaCl==1.5.0
pytz==2022.1
requests==2.26.0
